% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Utility to construct matrix of function values for coptourplot}
\usage{
app(x, y, f, par)
}
\arguments{
\item{x}{numeric. First input variable.}

\item{y}{numeric. Second input variable.}

\item{f}{expression. Gives the function to be evaluated.}

\item{par}{numeric.}
}
\value{
A matrix containing the function values of f for the inputs x and y.
}
\description{
Calculates a matrix containing the function values for a supplied function
and two input variables.
}
