% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{c_aic}
\alias{c_aic}
\title{Model selection criteria for fitcop objects: AIC}
\usage{
c_aic(copula)
}
\arguments{
\item{copula}{A list of class "fitcop". Supplies the value of the
log-likelihood function.}
}
\value{
The value of the AIC for the supplied copula and the related maximum
likelihood estimation.
}
\description{
Calculates AIC for a supplied fitted copula object.
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
U <- matrix(c(0.7026, 0.6359, 0.4116, 0.8833, 0.3127, 0.4035), nrow = 3,
byrow = TRUE)
fit <- cfit(copula = exc, data = U)
c_aic(fit)
}

}
\seealso{
\code{\link{cloglik}} and \code{\link{c_bic}}
}
