% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{cfit}
\alias{cfit}
\title{Fitting of copula parameter by maximum likelihood}
\usage{
cfit(copula, data, interval = NULL)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the
log-Likelihood function is to be maximised.}

\item{data}{Matrix or dataframe of appropriate dimension (n x d).
The data to base the Likelihood on. Data points have to be normed, i.e.
copula data has to lie within [0, 1]^d.}

\item{interval}{double. Optional argument, a vector to supply a lower and
upper bound of the interval to be searched for the maximum. If not supplied,
a very extensive interval within the appropriate parameter space will be
chosen. Plotting the log-Likelihood function to decide upon an interval is
recommended.}
}
\value{
A list of class "fitcop" inheriting the class of the supplied copula
object, with identical elements to the original object, aside from the
maximum likelihood estimate replacing the original parameter value. With one
additional element
\item{loglikelihood}{The maximum value of the log-likelihood of the supplied
copula.}
}
\description{
Calculates the maximum of the log-Likelihood function of a chosen Copula
family by utilizing stats::optimise.
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
U <- matrix(c(0.7026, 0.6359, 0.4116, 0.8833, 0.3127, 0.4035), nrow = 3,
byrow = TRUE)
fit <- cfit(copula = exc, data = U)
}

}
