% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clayCop.R
\name{claycop}
\alias{claycop}
\title{Construction of copula object for the Clayton family}
\usage{
claycop(par = NA, dim = 2L)
}
\arguments{
\item{par}{numeric. Supplies value of parameter of the copula to be
constructed.}

\item{dim}{integer. Supplies the number of dimensions of the copula to be
constructed.}
}
\value{
A list of class "claycop" with elements
\item{dimension}{Number of dimensions}
\item{generator}{List containing expressions for the generator function and
its inverse}
\item{parameter}{Parameter value of generator function}
\item{prange}{Vector of upper and lower bounds of permitted parameter values}
\item{family}{Name of constructed copula family, e.g. "Clayton"}
\item{distribution}{A list consisting of expressions for cdf and pdf}
The expressions can be evaluated for the parameter theta and additionally the
variable "t" in case of the generator function, "s" for its inverse and
variables "ui" in the case of cdf and pdf.
}
\description{
Constructs a claycop object, holding all relevant information on a given
copula of the "Clayton" family. A parameter called theta and the number of
dimensions has to be supplied. In this implementation Theta can only take
positive values. If the supplied parameter is too close to zero, you will be
informed that the use of an Independence copula is recommended in this case.
}
\examples{
\donttest{
excop <- claycop(par = 5, dim = 2)
summary(excop)
}

}
