% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLike.R
\name{CLike}
\alias{CLike}
\title{Likelihood of Copula}
\usage{
CLike(copula, marg)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the Likelhood function is to be calculated.}

\item{marg}{character. Supplied choices for the marginal distribution functions.
Can either be estimated through empirical functions or given by name.}
}
\value{
An expression of the calculated Likelihood function.
}
\description{
Calculates the Likelihood function of a chosen Copula family
}
\examples{
\donttest{
exC <- copu(Type = "Clayton", par = 5, dim = 2)
exS <- rCop(copula = exC, n = 1000)
}

}
