% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmethmo.R
\name{cmethmo}
\alias{cmethmo}
\title{Fitting of bivariate copula parameter by method of moments,
using Kendall's tau}
\usage{
cmethmo(copula, data = NULL, tau = NULL)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the
parameter is to be estimated.}

\item{data}{Matrix or dataframe of appropriate dimension (n x 2).
The data to base the dependence measure on. Optional if "tau" is supplied.}

\item{tau}{double. A value to take as Kendall's tau, to base the method of
moments estimator on.}
}
\value{
A copula object with parameter theta chosen by inversion of Kendall's
tau.
}
\description{
Calculates a method of moments estimate of the parameter of a chosen
bivariate Copula.
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
U <- matrix(c(0.7026, 0.6359, 0.4116, 0.8833, 3.2175, 4.0357), nrow = 3,
byrow = TRUE)
mmc <- cmethmo(copula = exc, data = U)
}

}
