% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppersplot.R
\name{coppersplot}
\alias{coppersplot}
\title{Perspective Plot of Copula}
\usage{
coppersplot(copula, FUN, n.grid = 26, delta = 0, xlim = 0:1,
  ylim = 0:1, zlim = NULL, xlab = "u1", ylab = "u2",
  theta = -45/2, phi = 30, expand = 0.75, col = c("red", "yellow"),
  shade = 0.2, ticktype = "detail",
  zlab = deparse(substitute(FUN))[1], ...)
}
\arguments{
\item{copula}{ArCop object. Supplies the copula to be used for the perspective plot.}

\item{FUN}{function. Gives the function to be plotted.}

\item{n.grid}{integer.}

\item{delta}{numeric.}

\item{xlim}{integer.}

\item{ylim}{integer.}

\item{xlab}{character.}

\item{ylab}{character.}

\item{nlevels}{integer.}
}
\value{
A perspective plot of a surface over the x-y plane using the values given by FUN for a copula of the form given by the supplied copula object.
}
\description{
Gives a perspective plot of a surface for a given copula object for a specified function
}
\examples{
\donttest{
exC <- copu(Type = "Clayton", par = 5, dim = 2)
exS <- rCop(copula = exC, n = 1000)
}

}
