% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coptourplot.R
\name{coptourplot}
\alias{coptourplot}
\title{Contourplot of Copula}
\usage{
coptourplot(copula, FUN, n.grid = 26, delta = 0, xlim = 0:1,
  ylim = 0:1, xlab = quote(u[1]), ylab = quote(u[2]))
}
\arguments{
\item{copula}{ArCop object. Supplies the copula to be used for the contourplot.}

\item{FUN}{function. Gives the function to be plotted}

\item{n.grid}{integer.}

\item{delta}{numeric.}

\item{xlim}{integer.}

\item{ylim}{integer.}

\item{xlab}{character.}

\item{ylab}{character.}
}
\value{
A contourplot with the vales given by FUN for a copula of the form given by the supplied copula object.
}
\description{
Gives a contourplot of a given copula object for a specified Function
}
\examples{
\donttest{
exC <- copu(Type = "Clayton", par = 5, dim = 2)
exS <- rCop(copula = exC, n = 1000)
}

}
