% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArCop.R
\name{copu}
\alias{copu}
\title{Construction of copula object}
\usage{
copu(Type = "Unknown", param = NA, dim = 2L, generator = NULL)
}
\arguments{
\item{Type}{character. Type can be supplied in order to construct a named copula. e.g. a Frank copula with type = "Frank".}

\item{param}{numeric. Supplies value of parameter(s) of the copula to be constructed.}

\item{dim}{interger. Supplies the number of dimensions of the copula to be constructed.}

\item{generator}{function. Supplies the}
}
\description{
Constructs an ArCop object, either by utilising Type for known archimedean copulas or generator for an arbitrary generator function.
}
\examples{
\donttest{
ExCop <- copu()
ExCop
}

}
