% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corken.R
\name{corken}
\alias{corken}
\title{Kendall's Tau (empirical)}
\usage{
corken(data, use = "everything", fast = FALSE)
}
\arguments{
\item{data}{Matrix or dataframe of appropriate dimension (n x d).
Supplies the data as a matrix or data frame.}

\item{use}{character. String (may be abbreviated) indicating the treatment of
missing values: "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs" (see \code{cor} for an explanation of the options).}

\item{fast}{logical. For large data sets without missing values, the
computation of Kendall's tau can be sped up considerably by using the
function \code{cor.fk} from the package \code{pcaPP}.}
}
\value{
Matrix with pairwise Kendall rank correlation coefficients between
the columns of the supplied data.
}
\description{
Wrapper function for computing all pairwise Kendall rank correlation
coefficients between the columns of a given data matrix.
}
\examples{
\donttest{
# standard usage
iris <- iris[, -5]
corken(iris)

# with missing values (fast not working)
iris[1,1] <- NA
corken(iris)
corken(iris, use = "complete.obs")
corken(iris, fast = TRUE)

# fast (only works without missing values)
corken(swiss, fast = TRUE)
}

}
