% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclayCop.R
\name{dcop.claycop}
\alias{dcop.claycop}
\title{Construction of Clayton pdf}
\usage{
\method{dcop}{claycop}(copula, eva, u)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{eva}{Logical. If eva = TRUE, the resulting expression will be evaluated
for the supplied arguments.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf,
only needed in case of eva = TRUE.}
}
\description{
Construction of Clayton pdf
}
