% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrankcop.R
\name{dcop.frankcop}
\alias{dcop.frankcop}
\title{Construction of Frank pdf}
\usage{
\method{dcop}{frankcop}(copula, eva, u)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{eva}{Logical. If eva = TRUE, the resulting expression will be evaluated
for the supplied arguments.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf,
only needed in case of eva = TRUE.}
}
\description{
Construction of Frank pdf
}
