% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindCop.R
\name{dCop.indCop}
\alias{dCop.indCop}
\title{Construction of Independence pdf}
\usage{
\method{dCop}{indCop}(copula, eva, u)
}
\arguments{
\item{copula}{ArCop object. Supplies the copula for which the pdf is to be constructed.}

\item{eva}{Logical. If eva = TRUE, the resulting expression will be evaluated for the supplied arguments.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf, only needed in case of eva = TRUE.}
}
\description{
Construction of Independence pdf
}
