% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCop.R
\name{dCop}
\alias{dCop}
\title{Density function of a Copula}
\usage{
dCop(copula)
}
\arguments{
\item{copula}{ArCop object. Supplies the copula for which the pdf is to be constructed.}
}
\value{
An expression for the pdf of the supplied copula.
}
\description{
Constructs the density function by differntiating the supplied cdf expression.
}
\examples{
\donttest{
exC <- copu(Type = "Clayton", par = 5, dim = 2)
exS <- rCop(copula = exC, n = 1000)
}

}
