% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debye.R
\name{debye}
\alias{debye}
\title{Utility to calculate the parameter theta of a Frank copula for a given value
of Kendall's tau}
\usage{
debye(x)
}
\arguments{
\item{x}{double. A numeric value as input for the Debye function of order one.}
}
\value{
The value of the Debye function of order one at x.
}
\description{
Simplistic implementation of the Debye function of order one. A
simplification of an implementation of Diethelm Wuertz in the package
"fCopulae".
}
