% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franktau.R
\name{franktau}
\alias{franktau}
\title{Utility to calculate the parameter theta of a Frank copula for a given value
of Kendall's tau}
\usage{
franktau(tau)
}
\arguments{
\item{tau}{double. A numeric value for Kendall's tau.}
}
\value{
A value for the parameter theta based on method of moments estimation
by the supplied Kendalls'tau.
}
\description{
Implementation of the relationship between theta and tau for a Frank copula.
A simplification of an implementation of Diethelm Wuertz in the package
"fCopulae". Original implementation falsely gives value of "Inf" for
tau = -1, this has been fixed in this implementation.
}
\references{
Frees and Valdez (1998). Understanding Relationships Using
Copulas. North American Actuarial Journal 2(1):1-25.
}
