% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcop.R
\name{indcop}
\alias{indcop}
\title{Construction of Independence copula object}
\usage{
indcop(dim = 2L)
}
\arguments{
\item{dim}{integer. Supplies the number of dimensions of the copula to be
constructed.}
}
\value{
A list of class "indcop" with elements
\item{dimension}{Number of dimensions}
\item{family}{Name of constructed copula family, here "Independence"}
\item{distribution}{A list consisting of expressions for cdf and pdf}
The expressions can be evaluated for the variables "ui" in the case of cdf
and pdf.
}
\description{
Constructs a indcop object, holding all relevant information on an
Independence copula. The number of dimensions has to be supplied.
}
\examples{
\donttest{
excop <- indcop(dim = 2)
summary(excop)
}

}
\references{
Hofert et al. (2018). Elements of Copula Modeling with R.
Springer.

Nelsen (2006). An introduction to copulas. Springer Series in Statistics.
Second Edition.

Frees and Valdez (1998). Understanding Relationships Using Copulas. North
American Actuarial Journal 2(1):1-25 January 1998.
}
\seealso{
\code{\link{claycop}} and \code{\link{frankcop}}
}
