% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvGenEval.R
\name{InvGenEval}
\alias{InvGenEval}
\title{Inverse Generator function utility}
\usage{
InvGenEval(s, copula)
}
\arguments{
\item{s}{numeric. Input for the inverted generator function to be applied.}

\item{copula}{ArCop object. Decides form of generator function.}
}
\value{
The value of the inverse generator function of the given ArCop object at input x.
}
\description{
Applies appropriate inverted generator function to input, based on supplied ArCop object.
}
