% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.claycop}
\alias{is.claycop}
\title{Utility to check inheritance of copula objects for "claycop"}
\usage{
is.claycop(copula)
}
\arguments{
\item{copula}{A copula object.}
}
\value{
TRUE or FALSE depending on whether the supplied copula object is of
the class "claycop".
}
\description{
Utility to check inheritance of copula objects for "claycop"
}
