% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.copula}
\alias{is.copula}
\title{Utility to check inheritance of objects for "copula"}
\usage{
is.copula(copula)
}
\arguments{
\item{copula}{A copula object.}
}
\value{
TRUE or FALSE depending on whether the supplied object is of
the class copula.
}
\description{
Utility to check inheritance of objects for "copula"
}
