% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.fitcop}
\alias{is.fitcop}
\title{Utility to check inheritance of objects for "fitcop"}
\usage{
is.fitcop(copula)
}
\arguments{
\item{copula}{A copula object.}
}
\value{
TRUE or FALSE depending on whether the supplied object is of
the class "fitcop".
}
\description{
Utility to check inheritance of objects for "fitcop"
}
