% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.frankCop}
\alias{is.frankCop}
\title{Utility to check inheritance of copula objects for "frankCop"}
\usage{
is.frankCop(copula)
}
\arguments{
\item{copula}{A copula object.}
}
\value{
TRUE or FALSE depending on whether the supplied copula object is of
the class "frankcop".
}
\description{
Utility to check inheritance of copula objects for "frankCop"
}
