% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfrankcop.R
\name{pcop.frankcop}
\alias{pcop.frankcop}
\title{Construction of Frank cdf}
\usage{
\method{pcop}{frankcop}(copula, u = NULL)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the cdf is to be
defined.}

\item{u}{double. Optional argument supplying the data to evaluate the cdf.}
}
\description{
Construction of Frank cdf
}
