% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pindCop.R
\name{pCop.indCop}
\alias{pCop.indCop}
\title{Construction of Independence cdf}
\usage{
\method{pCop}{indCop}(copula, eva = FALSE, u)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the cdf is to be defined.}

\item{eva}{Logical. If eva = TRUE, the resulting expression for the cdf will be evaluated for the supplied arguments.}

\item{u}{numeric. Optional argument supplying the data to evaluate the cdf, only needed in case of eva = TRUE.}
}
\description{
Construction of Independence cdf
}
