% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCop.R
\name{pCop}
\alias{pCop}
\title{Distribution function of a copula}
\usage{
pCop(copula)
}
\arguments{
\item{copula}{ArCop object. Supplies the copula for which the cdf is to be defined.}
}
\value{
An expression for the cdf of the supplied copula object.
}
\description{
Constructs the distribution function for a supplied copula object.
}
\examples{
\donttest{
exC <- copu(Type = "Clayton", par = 5, dim = 2)
exS <- rCop(copula = exC, n = 1000)
}

}
