% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.copula}
\alias{print.copula}
\title{Printing of copula objects}
\usage{
\method{print}{copula}(x, force = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{force}{logical. If "TRUE", the cdf and pdf of the copula object will be
printed despite their length.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing of copula objects
}
