% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fitcop}
\alias{print.fitcop}
\title{Printing of fitted copula objects}
\usage{
\method{print}{fitcop}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing of fitted copula objects
}
