% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rindcop.R
\name{rcop.indcop}
\alias{rcop.indcop}
\title{Generating random samples of an Independence copula}
\usage{
\method{rcop}{indcop}(copula = NULL, n = 0)
}
\arguments{
\item{copula}{A copula object. Decides the copula to sample from.}

\item{n}{integer. Supplies the desired number of observations.}
}
\description{
Generating random samples of an Independence copula
}
