% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcop.R
\name{rcop}
\alias{rcop}
\title{Sampling of Copula}
\usage{
rcop(copula = NULL, n = 0)
}
\arguments{
\item{copula}{A copula object. Decides the copula to sample from.}

\item{n}{integer. Supplies the desired number of observations.}
}
\value{
A matrix containing n observations of d random variables, based on
the family of the supplied copula object.
}
\description{
Constructs a sample based on a given copula object.
}
\examples{
\donttest{
excop <- Claycop(par = 5, dim = 2)
exsam <- rcop(copula = excop, n = 1000)
plot(exsam, ylim = c(0, 1), xlim = c(0, 1), ylab = "U2", xlab = "U1")
}

}
\references{
Hofert et al. (2018). Elements of Copula Modeling with R.
Springer.

Nelsen (2006). An introduction to copulas. Springer Series in Statistics.
Second Edition.

Whelan (2004). Sampling from Archimedean copulas. Quantitative
Finance, 2004, vol. 4, issue 3, pp. 339-352.

Marshall and Olkin (1988). Families of Multivariate Distributions.
Journal of the American Statistical Association, Vol. 83, No. 403, pp.834-841.

Genest (1987). Frank's family of bivariate distributions. Biometrika (1987),
74, 3, pp. 549-55.
}
\seealso{
\code{\link{pcop}} and \code{\link{dcop}}
}
