% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zofgrid.R
\name{zofgrid}
\alias{zofgrid}
\title{Utility to construct xyz-coordinates of a function for plotting functions}
\usage{
zofgrid(cop, FUN, n, delta, xlim, ylim)
}
\arguments{
\item{cop}{A copula object.}

\item{FUN}{character. Decides whether the cdf or pdf of the supplied copula
is to be plotted. Accepts "cdf" or "pdf" as input.}

\item{n}{integer. Number of grid points in x- and y-direction.}

\item{delta}{numeric. A value in the range [0, 0.5) for changing the
evaluation boundaries by adding it to the lower limits of x and y and
subtracting it from the upper limits of x and y.}

\item{xlim}{integer. Range of x values to be plotted.}

\item{ylim}{integer. Range of y values to be plotted.}
}
\value{
A list with a vector  of x-coordinates, a vector of y-coordinates and
a matrix with the function values of FUN on the xy-grid.
}
\description{
Sets up an xy-grid, has the supplied function evaluated on the grid, and
returns the resulting xyz-coordinates.
}
