% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfit.R
\name{cfit}
\alias{cfit}
\title{Fitting of copula parameter by maximum likelihood}
\usage{
cfit(copula, data, interval = NULL)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the
log-Likelihood function is to be maximised.}

\item{data}{Matrix or dataframe of appropriate dimension (n x d).
The data to base the Likelihood on. Data points have to be normed, i.e.
copula data has to lie within [0, 1]^d.}

\item{interval}{double. Optional argument, a vector to supply a lower and
upper bound of the interval to be searched for the maximum.}
}
\value{
A list containing a copula object with parameter theta chosen by
maximum Likelihood estimation and the value of the log-Likelihood function at
the found optimum.
}
\description{
Calculates the maximum of the log-Likelihood function of a chosen Copula
family by utilizing stats::optimise.
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
U <- matrix(c(0.7026, 0.6359, 0.4116, 0.8833), nrow = 2, byrow = TRUE)
ll <- cloglik(copula = exc, data = U)
}

}
