% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clayCop.R
\name{clayCop}
\alias{clayCop}
\title{Construction of copula object for the Clayton family}
\usage{
clayCop(par = NA, dim = 2L)
}
\arguments{
\item{par}{numeric. Supplies value of parameter of the copula to be constructed.}

\item{dim}{interger. Supplies the number of dimensions of the copula to be constructed.}
}
\value{
A list of class "clayCop" with elements
\item{dimension}{Number of dimensions}
\item{expression}{List containing expression for generator function and its inverse}
\item{parameter}{Parameter value of generator function}
\item{prange}{vector of upper and lower bounds of the parameter}
\item{family}{Name of constructed copula, e.g. "Clayton"}
\item{distribution}{A list consisting of cdf and pdf}
}
\description{
Constructs a clayCop object, holding all relevant information on a given
copula of the "Clayton" family. A parameter theta and the number of
dimension has to be supplied. Theta can only take positive values.
If the supplied parameter is too close to zero, a warning will be given;
the use of an Independence copula is recommended in this case.
}
\examples{
\donttest{
exCop <- clayCop(par = 5, dim = 2)
summary(exCop)
}

}
