% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clik.R
\name{clik}
\alias{clik}
\title{Likelihood of Copula}
\usage{
clik(copula, marg)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the
Likelhood function is to be calculated.}

\item{marg}{character. Supplies choices for the marginal distribution
functions.
Can either be estimated through empirical functions or given by name.}
}
\value{
An expression of the calculated Likelihood function.
}
\description{
Calculates the Likelihood function of a chosen Copula family
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
exs <- rcop(copula = exc, n = 1000)
}

}
