% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clik.R
\name{cloglik}
\alias{cloglik}
\title{Likelihood of Copula}
\usage{
cloglik(copula, data, parameter = NULL)
}
\arguments{
\item{copula}{A copula object. Decides the copula family for which the
Likelhood function is to be calculated.}

\item{data}{The data to base the Likelihood on.}

\item{parameter}{Parameter value of the copula.}
}
\value{
The value of the calculated loglikelihood function.
}
\description{
Calculates the Likelihood function of a chosen Copula family
}
\examples{
\donttest{
exc <- claycop(par = 5, dim = 2)
U <- matrix(c(0.7026, 0.6359, 0.4116, 0.8833), nrow = 2, byrow = TRUE)
ll <- cloglik(copula = exc, data = U)
}

}
