% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppersplot.R
\name{coppersplot}
\alias{coppersplot}
\title{Perspective Plot of Copula}
\usage{
coppersplot(copula, FUN, n.grid = 26, delta = 0, xlim = 0:1,
  ylim = 0:1, xlab = "u1", ylab = "u2", zlab = FUN,
  main = paste(FUN, "of", copula$family, "Copula", sep = " "),
  theta = -45/2, phi = 30, expand = 0.75, col.pal = "Viridis",
  border = NA, shade = NA, ticktype = "detail", ...)
}
\arguments{
\item{copula}{A copula object. Supplies the copula to be used for the
perspective plot.}

\item{FUN}{character. Decides whether the cdf or pdf of the supplied copula
is to be plotted. Accepts "cdf" or "pdf" as input.}

\item{n.grid}{integer. Number of grid points in x- and y-direction. Vector of
length two or one (used for both directions). Determines the number of facets.}

\item{delta}{double. A value in the range [0, 0.5) for changing the
evaluation boundaries by adding it to the lower limits of x and y and
subtracting it from the upper limits of x and y. Defaults to zero.}

\item{xlim}{double. Range of x values to be plotted. Defaults to [0, 1].}

\item{ylim}{double. Range of y values to be plotted. Defaults to [0, 1].}

\item{xlab}{character. Title for x-axis.}

\item{ylab}{character. Title for y-axis.}

\item{zlab}{character. Title for z-axis.}

\item{main}{character. Main title for the plot. Set to NULL to omit the title.}

\item{theta}{double. Azimuthal viewing direction.}

\item{phi}{double. Viewing colatitude.}

\item{expand}{double. Expansion factor applied to the z coordinates.}

\item{col.pal}{character. Name of the color palette to be used for the sur-
face facets, provided by the function hcl.colors() in the grDevices package.}

\item{border}{character, function, or integer. Color of the line drawn around
the surface facets. The default NA shows no borders. NULL corresponds to
par("fg").}

\item{shade}{double. Shade of surface facets.}

\item{ticktype}{character. "simple" draws just an arrow, "detailed" draws
normal ticks as in 2D plots.}

\item{...}{character, function, or integer. Additional graphical parameters.}
}
\value{
A perspective plot of a surface over the x-y plane using the values
given by FUN for a copula of the form given by the supplied copula object.
}
\description{
Gives a perspective plot of the surface of a given copula object for a
specified function.
}
\examples{
\donttest{
excop <- claycop(par = 1, dim = 2)
# Plotting the cdf
coppersplot(copula = excop, FUN = "cdf")
# Plotting the pdf
coppersplot(copula = excop, FUN = "pdf", n.grid = 20, delta= 0.1,
            theta = -30, col.pal = "Plasma", border = NULL,
            main = "Clayton Copula PDF", cex.main = 0.8)
}

}
