% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coptourplot.R
\name{coptourplot}
\alias{coptourplot}
\title{Contourplot of Copula}
\usage{
coptourplot(copula, FUN, n.grid = 50, delta = 0, xlim = 0:1,
  ylim = 0:1, xlab = quote(u[1]), ylab = quote(u[2]),
  main = paste(FUN, "of", copula$family, "Copula", sep = " "),
  nlevels = 20, col.pal = "Viridis", asp = 1, con.lines = TRUE,
  ...)
}
\arguments{
\item{copula}{A copula object. Supplies the copula to be used for the
contourplot.}

\item{FUN}{character. Decides whether the cdf or pdf of the supplied copula
is to be plotted. Accepts "cdf" or "pdf" as input.}

\item{n.grid}{integer. Number of grid points in x- and y-direction. Vector of
length two or one (used for both directions).}

\item{delta}{double. A value in the range [0, 0.5) for changing the
evaluation boundaries by adding it to the lower limits of x and y and
subtracting it from the upper limits of x and y. Defaults to zero.}

\item{xlim}{double. Range of x values to be plotted. Defaults to [0, 1].}

\item{ylim}{double. Range of y values to be plotted. Defaults to [0, 1].}

\item{xlab}{character. Title for x-axis.}

\item{ylab}{character. Title for y-axis.}

\item{main}{character. Title for the main plot. Set to NULL to omit the title.}

\item{nlevels}{integer. Number of contour levels to be plotted.}

\item{col.pal}{character. Name of the color palette to be used for the contour
levels, provided by the function hcl.colors() in the grDevices package.}

\item{asp}{double. Aspect ratio of the axes. Defaults to 1. Use NA to fit
the aspect ratio to the available drawing area.}

\item{con.lines}{logical. Draw contour lines if TRUE (default).}

\item{...}{character, function, or integer. Additional graphical parameters.}
}
\value{
A contourplot with the values given by FUN for a copula of the form
given by the supplied copula object.
}
\description{
Gives a contourplot of a given copula object for a specified function.
}
\examples{
\donttest{
excop <- claycop(par = 0.5, dim = 2)
# Plotting the cdf
coptourplot(copula = excop, FUN = "cdf")
# Plotting the pdf
coptourplot(copula = excop, FUN = "pdf", n.grid = 42, nlevels = 16,
            col.pal = "YlGnBu",
            plot.title = title(main = "Clayton Copula PDF"))
}

}
