% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coPtourplot.R
\name{coPtourplot}
\alias{coPtourplot}
\title{Contourplot of Copula}
\usage{
coPtourplot(copula, FUN, n.grid = 26, delta = 0, nlevels = 20,
  xlim = 0:1, ylim = 0:1, xlab = quote(u[1]), ylab = quote(u[2]))
}
\arguments{
\item{copula}{A copula object. Supplies the copula to be used for the
contourplot.}

\item{FUN}{character. Decides whether the cdf or pdf of the supplied copula
is to be plotted. Accepts "cdf" or "pdf" as input.}

\item{n.grid}{integer.}

\item{delta}{numeric.}

\item{nlevels}{integer.}

\item{xlim}{integer.}

\item{ylim}{integer.}

\item{xlab}{character.}

\item{ylab}{character.}
}
\value{
A contourplot with the values given by FUN for a copula of the form
given by the supplied copula object.
}
\description{
Gives a contourplot of a given copula object for a specified function.
}
\examples{
\donttest{
exCop <- clayCop(par = 5, dim = 2)
# Plotting the cdf
coPtourplot(copula = exCop, FUN = "cdf", n.grid = 26, delta = 0, nlevels = 20,
xlim = 0:1, ylim = 0:1, xlab = quote(u[1]), ylab = quote(u[2]))
# Plotting the pdf
coPtourplot(copula = exCop, FUN = "pdf", n.grid = 26, delta = 0, nlevels = 20,
xlim = 0:1, ylim = 0:1, xlab = quote(u[1]), ylab = quote(u[2]))
}

}
