% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclaycop.R
\name{dcop.claycop}
\alias{dcop.claycop}
\title{Construction of Clayton pdf}
\usage{
\method{dcop}{claycop}(copula, u = NULL)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf.}
}
\description{
Construction of Clayton pdf
}
