% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dclayCop.R
\name{dCop.clayCop}
\alias{dCop.clayCop}
\title{Construction of Clayton pdf}
\usage{
\method{dCop}{clayCop}(copula, eva, u)
}
\arguments{
\item{copula}{ArCop object. Supplies the copula for which the pdf is to be constructed.}

\item{eva}{Logical. If eva = TRUE, the resulting expression will be evaluated for the supplied arguments.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf, only needed in case of eva = TRUE.}
}
\description{
Construction of Clayton pdf
}
