% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrankcop.R
\name{dcop.frankcop}
\alias{dcop.frankcop}
\title{Construction of Frank pdf}
\usage{
\method{dcop}{frankcop}(copula, u = NULL)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{u}{numeric. Optional argument supplying the data to evaluate the pdf.}
}
\description{
Construction of Frank pdf
}
