% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindcop.R
\name{dcop.indcop}
\alias{dcop.indcop}
\title{Construction of Independence pdf}
\usage{
\method{dcop}{indcop}(copula, u = NULL)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{u}{double. Optional argument supplying the data to evaluate the pdf.}
}
\description{
Construction of Independence pdf
}
