% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dcop}
\alias{dcop}
\title{Density function of a Copula}
\usage{
dcop(copula, u)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the pdf is to be
constructed.}

\item{u}{double. Optional argument supplying the data to evaluate the pdf.}
}
\value{
expression or numeric for the pdf of the supplied copula object or
its value at the supplied u.
}
\description{
Constructs the density function by differntiating the supplied cdf expression.
}
\examples{
\donttest{
excop <- claycop(par = 5, dim = 2)
expdf <- dcop(excop, eva = FALSE)
#In case evaluation of the pdf is wanted:
expdfval <- dcop(excop, eva = TRUE, u = c(0.5, 0.5))
}

}
\references{
Hofert et al. (2018). Elements of Copula Modeling with R.
Springer.

Nelsen (2006). An introduction to copulas. Springer Series in Statistics.
Second Edition.
}
\seealso{
\code{\link{pcop}} and \code{\link{rcop}}
}
