% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankCop.R
\name{frankCop}
\alias{frankCop}
\title{Construction of copula object for the Frank family}
\usage{
frankCop(par = NA, dim = 2L)
}
\arguments{
\item{par}{numeric. Supplies value of parameter of the copula to be constructed.}

\item{dim}{interger. Supplies the number of dimensions of the copula to be constructed.}
}
\value{
A list of class "frankCop" with elements
\item{dimension}{Number of dimensions}
\item{expression}{List containing expression for generator function and its inverse}
\item{parameter}{Parameter value of generator function}
\item{prange}{vector of upper and lower bounds of the parameter}
\item{family}{Name of constructed copula, e.g. "Frank"}
\item{distribution}{A list consisting of cdf and pdf}
}
\description{
Constructs a frankCop object.
}
\examples{
\donttest{
exCop <- frankCop(par = 5, dim = 2)
exCop
}

}
