% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indCop.R
\name{indCop}
\alias{indCop}
\title{Construction of Independence copula object}
\usage{
indCop(dim = 2L)
}
\arguments{
\item{dim}{integer. Supplies the number of dimensions of the copula to be
constructed.}
}
\value{
A list of class "indCop" with elements
\item{dimension}{Number of dimensions}
\item{family}{Name of constructed copula family, e.g. "Clayton"}
\item{distribution}{A list consisting of expressions for cdf and pdf}
The expressions can be evaluated for the variables "ui" in the case of cdf
and pdf.
}
\description{
Constructs a indCop object, holding all relevant information on an
Independence copula. The number of dimensions has to be supplied.
}
\examples{
\donttest{
exCop <- indCop(dim = 2)
summary(exCop)
}

}
