% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invGenEval.R
\name{invGenEval}
\alias{invGenEval}
\title{Inverse Generator function utility}
\usage{
invGenEval(s, copula)
}
\arguments{
\item{s}{numeric. Input for the inverted generator function to be applied.}

\item{copula}{A copula object. Decides form of generator function.}
}
\value{
The value of the inverse generator function of the given copula object at input x.
}
\description{
Applies appropriate inverted generator function to input, based on supplied copula object.
}
