% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.indcop}
\alias{is.indcop}
\title{Utility to check inheritance of copula objects for "indcop"}
\usage{
is.indcop(copula)
}
\arguments{
\item{copula}{A copula object.}
}
\value{
TRUE or FALSE depending on whether the supplied copula object is of
the class "indcop".
}
\description{
Utility to check inheritance of copula objects for "indcop"
}
