% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclaycop.R
\name{pcop.claycop}
\alias{pcop.claycop}
\title{Construction of Clayton cdf}
\usage{
\method{pcop}{claycop}(copula, u = NULL)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the cdf is to be
defined.}

\item{u}{double. Optional argument supplying the data to evaluate the cdf.}
}
\description{
Construction of Clayton cdf
}
