% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcop.R
\name{pcop}
\alias{pcop}
\title{Distribution function of a copula}
\usage{
pcop(copula, u)
}
\arguments{
\item{copula}{A copula object. Supplies the copula for which the cdf is to be
defined.}

\item{u}{numeric. Optional argument supplying the data to evaluate the cdf.}
}
\value{
An expression for the cdf of the supplied copula object or its
numeric value at the supplied u in case evaluation was chosen..
}
\description{
Constructs the distribution function for a supplied copula object.
}
\examples{
\donttest{
excop <- Claycop(par = 5, dim = 2)
excdf <- pcop(excop)
#In case evaluation of the cdf is wanted:
excdfval <- pcop(excop, u = c(0.5, 0.5))
}

}
