% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rClayCop.R
\name{rCop.clayCop}
\alias{rCop.clayCop}
\title{Generating random samples of a Clayton copula based on the algorithm of
Marshall and Olkin (1988)}
\usage{
\method{rCop}{clayCop}(copula = NULL, n = 0)
}
\arguments{
\item{copula}{A copula object. Decides the copula to sample from.}

\item{n}{integer. Supplies the desired number of observations.}
}
\description{
Generating random samples of a Clayton copula based on the algorithm of
Marshall and Olkin (1988)
}
