% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCop.R
\name{rcop}
\alias{rcop}
\title{Sampling of Copula}
\usage{
rcop(copula = NULL, n = 0)
}
\arguments{
\item{copula}{A copula object. Decides the copula to sample from.}

\item{n}{integer. Supplies the desired number of observations.}
}
\value{
A matrix containing n observations of d random variables, based on
the family of the supplied copula object.
}
\description{
Constructs a sample based on a given copula object.
}
\examples{
\donttest{
excop <- Claycop(par = 5, dim = 2)
exsam <- rcop(copula = excop, n = 1000)
plot(exsam, ylim = c(0, 1), xlim = c(0, 1), ylab = "U2", xlab = "U1")
}

}
