% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.copula}
\alias{summary.copula}
\title{Summary of copula objects}
\usage{
\method{summary}{copula}(x, forced = FALSE)
}
\arguments{
\item{x}{a result of the \emph{default} method of \code{summary()}.}

\item{force}{logical. If "TRUE", the cdf and pdf of the copula object will be
included despite their length.}
}
\description{
Summary of copula objects
}
